OPEN - task 13: Logging configuration https://github.com/paypal/sdk-core-java/issues/issue/13
This is to address an issue where the logger name of multiple classes is
being mangled to some degree.

Normally, when using logging frameworks like log4j, logback, or slf4j,
they provide a factory method that takes a java.util.Class<?> argument.

For broadest compatibility however, the paypal APIs are built on top of
the java.util.logging framework. Logging is funnelled through the
"Logging Manager" which is cataloging loggers for each individual class.

The logger used for each class is determined by an argument to a method
call.  The class argument passed, is then converted using ".toString()"
which results in "class <class-canonical-name".  

As suggested by another github user, this should just be
.getCanonicalName() instead of .toString()